+++ packages/core/src/page/conditions/DialogCondition.spec.ts
@@ -27,6 +27,7 @@ describe('Condition', () => {
 			page.click('#alert')
 
 			const alert: Dialog = await condition.waitForEvent(page)
+			expect(alert).not.toBeNull()
 			expect(alert.message()).toBe('ok')
 			await alert.dismiss()
 		})
@@ -36,19 +37,24 @@ describe('Condition', () => {
 
 			page.click('#confirm')
 
-			const alert: Dialog = await condition.waitForEvent(page)
+			const alert: Dialog | null = await condition.waitForEvent(page)
+			expect(alert).not.toBeNull()
 			expect(alert.message()).toBe('set the value')
 			await alert.accept()
 		})
 
 		test('waits Until.alertIsPresent prompt', async () => {
 			const condition = Until.alertIsPresent()
+			await page.waitForSelector('#prompt', { visible: true })
 
-			page.click('#prompt')
+			page.click('#prompt', { delay: 100 })
 
-			const alert: Dialog = await condition.waitForEvent(page)
+			const alert = await condition.waitForEvent(page)
+			expect(alert).not.toBeNull()
 			expect(alert.message()).toBe('enter your name')
 			await alert.accept('Ivan')
+
+			await page.waitForSelector('#prompt', { visible: true })
 		})
 	})
 })