+++ packages/core/src/runtime/Browser.ts
@@ -18,7 +18,7 @@ import { PuppeteerClientLike } from '../driver/Puppeteer'
 import { WorkRoot } from '../runtime-environment/types'
 import KSUID from 'ksuid'
 import { Key } from '../page/Enums'
-import termImg from 'term-img'
+// import termImg from 'term-img'
 import { ConcreteTestSettings } from './Settings'
 import { NetworkErrorData, ActionErrorData } from './errors/Types'
 import { StructuredError } from '../utils/StructuredError'
@@ -127,7 +127,7 @@ export class Browser<T> implements BrowserInterface {
 	}
 
 	@addCallbacks()
-	public async wait(timeoutOrCondition: Condition | number): Promise<boolean> {
+	public async wait(timeoutOrCondition: Condition | number): Promise<any> {
 		if (typeof timeoutOrCondition === 'number') {
 			await new Promise(yeah => setTimeout(yeah, Number(timeoutOrCondition) * 1e3))
 			return true
@@ -540,12 +540,12 @@ export class Browser<T> implements BrowserInterface {
 			this.screenshots.push(path)
 			debugScreenshot(`Saved screenshot to: ${path}`)
 
-			termImg(path, {
-				width: '40%',
-				fallback: () => {
-					return `Screenshot path: ${path}`
-				},
-			})
+			// termImg(path, {
+			// 	width: '40%',
+			// 	fallback: () => {
+			// 		return `Screenshot path: ${path}`
+			// 	},
+			// })
 		}
 	}
 }